...G...PS. [ZBX-7288] removed unused variable, fixed compiler warning
git-svn-id: svn://svn.zabbix.com/branches/dev@40873 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-7288/src/zabbix_proxy/proxyconfig/proxyconfig.c
@@ -77,7 +77,7 @@ static void	process_configuration_sync(size_t *data_size)
 		goto out;
 	}
 
-	*data_size = jp.end - jp.start + 1;     /* performance metric */
+	*data_size = (size_t)(jp.end - jp.start + 1);     /* performance metric */
 
 	/* if the answer is short then most likely it is a negative answer "response":"failed" */
 	if (128 > *data_size &&
